Interrupt ID propagation to S-EL1
The goal is to changes the way of interrupt handling and particularly the
acknowledgment. By this patch the interrupt acknowledge is done in EL3 dispatcher while the
interrupt end in the S-EL1 payload.
This interrupt handling approach enables handling all Secure
interrupts in EL3 and if the interrupt is dedicated to EL3 only to acknowledge
and end it in EL3 dispatcher. The bottleneck is in adding interrupt id in of
EL3->SEL1 communication.

This patch is extending the protocol between the EL3 dispatcher and S-EL1
payload by passing the interrupt ID as third argument (register: X2).
The changes are in 'tspd_sel1_interrupt_handler' where every handled interrupt
is acknowledged and the ID is passed to 'tsp_fiq_handler'. The ID passing changes
the 'tsp_update_sync_fiq_stats' such it has to retransmit the ID to
'tsp_fiq_handler'. In order to avoid double acknowledge in 'tsp_fiq_handler', an
protocol is introduced allowing the 'tsp_fiq_handler' to distinguish between
interrupts handled in EL3 and in S-EL1 (those handled in S-EL1 has to be
acknowledged).

Signed-off-by: Petar Lalov <p.lalov@virtualopensystems.com>
Signed-off-by: Kevin Chappuis <k.chappuis@virtualopensystems.com>