mt8173: Add buck plat parameters and memory configuration for BL31
We add plat parameter structs to support BL2 to pass variable-length,
variable-type parameters to BL31. The parameters are structured as a
link list. During bl31 setup time, we travse the list to process each
parameter. If the parameter is for certain hardware, such as buck,
we will initialize it and register the callback function. Later on,
soc can control the hardware via the callback functions.

Memory configuration is specified in build command by specifying the
MEDIATEK_MEM_CONFIG. The default configuration is SRAM_192K, which
define the memory layout for BL31 to run on 192KB SRAM.

Change-Id: I923628d7296c84fb8179c5355b4cfd5dc9a9e786
Signed-off-by: Jimmy Huang <jimmy.huang@mediatek.com>